Sema: Fix inconsistent diagnostics on ambiguous access-levels
The call on `findInconsistentImportsAcrossModule` used the explicit
import as anchor for the check. However since #74035 we skip reporting
ambiguity with public imports. So if the anchor was the public import we
may ignore an implicit import that would otherwise match against a
different non-public explicit import. Invert the predicate for the
diagnostic to be reliably raised as expected, no matter the file
ordering.

rdar://144615496