[cxx-interop] Import non-public members of SWIFT_PRIVATE_FILEID-annotated classes
The ImportNonPublicCxxMembers experimental feature flag was introduced
to gate the importation of non-public class members, to ward off latent
bugs in projects that have no need for importing private members.
However that flag is required by the SWIFT_PRIVATE_FILEID feature to
access said private fields.

This patch relaxes the "don't import this non-public member" condition
a bit by bypassing this feature flag for non-public members of C++
classes that have the SWIFT_PRIVATE_FILEID annotation, thus allowing
that feature to be used without specifying the experimental feature
flag. Since that annotation is opt-in by design, ClangImporter will
still skip over the private members of most C++ classes (when
ImportNonPublicCxxMembers is not enabled), so that projects not using
this feature are unaffected.

ImportNonPublicCxxMembers is still useful (i.e., should be enabled by
default in the future) because it improves the diagnostics for when
a user tries to access any non-public member in Swift.

rdar://146304086