[Sema] Dig out other constructor if call is wrapped in a Dot-Call
Guessing this bug /feature is introduced in CSApply ExprRewriter::buildMemberRef
but didn't dig deep enough to find there. If a derived class' designated initializer
initializes all stored let properties before calling super.init(), we don't properly
type-check the chained init call. This does not happen if the stored properties are
var or default initialized. We can however dig the OtherConstructorDeclRefExpr out of
the semanticFn of the DotSyntaxCallExpr and proceed as normal.