Sema: Ban global actor on getters in a future lang mode
Global actor attributes were hitherto allowed on getters and (until
Swift 6) other accessors, but intendedly ignored in favor of the
enclosing `var` or `subscript` declaration's isolation when accessing
the storage declaration. This change deprecates the getter case as well,
to become an error in the next language mode.

Also do a better job at suggesting to move the attribute to the storage:
* Tighten up the condition for when the attr is legal on the storage.
* Don't suggest if the storage already specifies some kind of explicit
  isolation.
* Highlight the attr.
* Also remove the attr from the accessor.