[cxx-interop] Fix access check for nested private C++ enums
This patch fixes the access check for nested private C++ enums to look
for the SWIFT_PRIVATE_FILEID of the enclosing C++ class, if any.
Previously, the check was looking at for SWIFT_PRIVATE_FILEID on the
enum decl itself (which is meaningless); that prevented nested private
enum members from being accessible in Swift.

This patch also specializes the type signature of getPrivateFileIDAttrs
to clarify the fact that SWIFT_PRIVATE_FILEID is not a meaningful
annotation on anything other than CXXRecordDecl, because that is the
only kind of decl that can assign access specifiers to its members.

rdar://148081340