Prepare diagnostics docs for docc generation
Update the diagostic docs so that they can be built with docc:
    * Add technologyroot (diagnostics.md) to aggregate all markdown files
    * Fixed warnings in a few diagnostics docs

This allows the creation of docc documentation with a single
invocation of docc. It will generate the landing page from
diagnostics.md and include all of the diagnostics as subarticles.

The subarticles are accessible via the name of the markdown file,
which is being used as the ID for these diagnostics.
E.g. existential-any.

The docc docs for these documents can be generated locally with the
following command:
>docc preview --allow-arbitrary-catalog-directories userdocs/diagnostics