[CSRanking] Add a special ranking rule for SwiftUI's `overlay` and `background`
The disfavored overloads of `.overlay(_:alignment)` and `.background(_:alignment)`
are always strictly worse for these methods. Other modifiers are overloaded on
result i.e. `some View` vs. `some ShapeStyle` so introducing `@_disfavoredOverload`
on them has source compatibility impact when they passed as an argument to
`overlay` or `background` because they disfavor version that accepts a type
that conforms to `View` over the one that accepts a `ShapeStyle`.