[Runtimes][CMake] Set up supplemental super-build
As we add libraries, it will become more tedious to manage building
all of them independently. When built together, they should mostly take
the same arguments specifying where they are installed, what compilers
to use, and what configuration to use when building them. The
super-build pattern is useful for this purpose. For certain
distributions, it is useful to configure libraries with specific flags,
but for most purposes, that is unnecessary.

This hooks up a super-build pattern and enables building the
Differentiation library as part of CI.